% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataThesis.R
\docType{data}
\name{coef.dataThesis}
\alias{coef.dataThesis}
\title{Coefficients used in simulated data set generated with fngendata}
\format{
A vector containing the coefficients used to generate dataThesis
}
\usage{
coef.dataThesis
}
\description{
Coefficients used in simulated data, used to demonstrate the functions of GCEstim.
}
\examples{

coef.dataThesis
}
\keyword{datasets}
