% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmgceclassfunctions.R
\name{coef.lmgce}
\alias{coef.lmgce}
\title{Extract \code{\link{lmgce}} Model Coefficients}
\usage{
\method{coef}{lmgce}(object, ...)
}
\arguments{
\item{object}{Fitted \code{\link{lmgce}} model object.}

\item{...}{Additional arguments (not used).}
}
\value{
Returns the coefficients from a \code{\link{lmgce}} object
}
\description{
Extract coefficients from a \code{\link{lmgce}} object
}
\examples{
\donttest{
res_gce_package <-
  lmgce(y ~ .,
        data = dataGCE,
        boot.B = 50,
        seed = 230676)
}
coef(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
