% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.tsbootgceclassfunctions.R
\name{coefficients.cv.tsbootgce}
\alias{coefficients.cv.tsbootgce}
\title{Extract \code{\link{cv.tsbootgce}} Model Coefficients}
\usage{
\method{coefficients}{cv.tsbootgce}(object, which = NULL, OLS = FALSE, seed = object$seed, ...)
}
\arguments{
\item{object}{Fitted \code{\link{cv.tsbootgce}} model object.}

\item{which}{The default is \code{which = NULL} and returns the coefficients
defined in the argument \code{coef.method} from the \code{cv.tsbootgce}
object. Can be set as "mode" or "median" and the mode and median coefficients
will be computed, respectively (see \code{\link[hdrcde]{hdr}}).}

\item{OLS}{Boolean value. If \code{TRUE}, the confidence interval returned
is based on OLS estimates. The default is \code{OLS = FALSE}.}

\item{seed}{A single value, interpreted as an integer, for reproducibility
or \code{NULL} for randomness. The default is \code{seed = object$seed}.}

\item{...}{Additional arguments.}
}
\value{
Returns the coefficients from a \code{cv.tsbootgce} object
}
\description{
Extract coefficients from a \code{\link{cv.tsbootgce}} object
}
\examples{
\donttest{
res.cv.tsbootgce <-
  cv.tsbootgce(
    formula = CO2 ~ 1 + L(GDP, 1) + L(EPC, 1) + L(EU, 1),
    data = moz_ts)

coefficients(res.cv.tsbootgce)
}

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
