% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGCE.test.R
\docType{data}
\name{dataGCE.test}
\alias{dataGCE.test}
\title{Simulated data set generated with fngendata}
\format{
A \code{data.frame} containing:
\describe{
  \item{X001}{A N(0,1) independent variable.}
  \item{X002}{A N(0,1) independent variable.}
  \item{X003}{A N(0,1) independent variable.}
  \item{X004}{A N(0,1) independent variable.}
  \item{X005}{A N(0,1) independent variable.}
  \item{y}{A Dependent variable: y = 1 + 3 * X003 + 6 * X004 + 9 * X005 + error;
  the error follows a normal distribution with mean equal to zero and variance
  such that the signal to noise ratio is equal to 5.}
}
}
\usage{
dataGCE.test
}
\description{
Simulated data, used to demonstrate the functions of GCEstim. The seed used
is the different from the one used to generate \code{dataGCE} but the
remaining parameters are the same.
}
\examples{

data(dataGCE.test)

plot(dataGCE.test)
}
\keyword{datasets}
