% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataincRidGME.R
\docType{data}
\name{dataincRidGME}
\alias{dataincRidGME}
\title{Simulated data set generated with fngendata}
\format{
A \code{data.frame} containing:
\describe{
  \item{X001}{A N(0,1) independent variable.}
  \item{X002}{A N(0,1) independent variable.}
  \item{X003}{A N(0,1) independent variable.}
  \item{X004}{A N(0,1) independent variable.}
  \item{X005}{A N(0,1) independent variable.}
  \item{X006}{A N(0,1) independent variable.}
  \item{y}{A Dependent variable: y = 2.5 - 8 * X004 + 19 * X005 -
  13 * X006 + error; the error follows a normal distribution with mean equal
  to zero and variance such that the signal to noise ratio is equal to 1.}
}
}
\usage{
dataincRidGME
}
\description{
Simulated data, used to demonstrate the functions of GCEstim.
}
\examples{

data(dataincRidGME)

plot(dataincRidGME)
}
\keyword{datasets}
