% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmgceclassfunctions.R
\name{df.residual.lmgce}
\alias{df.residual.lmgce}
\title{Residual Degrees-of-Freedom}
\usage{
\method{df.residual}{lmgce}(object, ...)
}
\arguments{
\item{object}{Fitted \code{\link{lmgce}} model object.}

\item{...}{additional arguments.}
}
\value{
The value of the residual degrees-of-freedom extracted from a
\code{\link{lmgce}} object.
}
\description{
Returns the residual degrees-of-freedom extracted from a fitted model
\code{\link{lmgce}} object.
}
\examples{
\donttest{
res_gce_package <-
  lmgce(y ~ .,
        data = dataGCE,
        boot.B = 50,
        seed = 230676)
}
df.residual(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
