% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ECT.test}
\alias{ECT.test}
\title{Entropy Concentration Theorem test}
\usage{
ECT.test(object, level = 0.95)
}
\arguments{
\item{object}{fitted \code{\link{lmgce}} object.}

\item{level}{the confidence level required. The default is
\code{level = 0.95}.}
}
\description{
The Entropy Concentration Theorem
}
\examples{
res_gce_package <-
  lmgce(y ~ .,
        data = dataGCE,
        seed = 230676)

ECT.test(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
