% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmgceclassfunctions.R
\name{formula.lmgce}
\alias{formula.lmgce}
\title{Extract Model Formula from \code{\link{lmgce}} object}
\usage{
\method{formula}{lmgce}(x, ...)
}
\arguments{
\item{x}{fitted \code{\link{lmgce}} object.}

\item{...}{additional arguments.}
}
\value{
An object of class \code{formula} representing the model formula.
}
\description{
Returns the model used to fit \code{\link{lmgce}} object.
}
\examples{
\donttest{
res_gce_package <-
  lmgce(y ~ .,
        data = dataGCE,
        boot.B = 50,
        seed = 230676)

formula(res_gce_package)
}

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
