% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmgceAddin.R
\name{lmgceAddin}
\alias{lmgceAddin}
\title{An add-in to easily generate the code for a \code{\link{lmgce}}
analysis}
\usage{
lmgceAddin()
}
\value{
The code to be use in the \code{\link{lmgce}} analysis.
}
\description{
Select data and choose the arguments to be used. The
execution of the code is also possible within the addin.
}
\details{
An addin for \code{\link{lmgce}}
}
\examples{
\donttest{
lmgceAddin()
}
}
