% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmgceAPP.R
\name{lmgceAPP}
\alias{lmgceAPP}
\title{\code{\link{lmgce}} Shiny application}
\usage{
lmgceAPP()
}
\value{
\code{NULL}. This function is called for its side effect
(launching the app).
}
\description{
A Shiny application to execute \code{\link{lmgce}}
}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
