% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgetraceclassfunctions.R
\name{plot.ridgetrace}
\alias{plot.ridgetrace}
\title{Plot Diagnostics for a \code{\link{ridgetrace}} Object}
\usage{
\method{plot}{ridgetrace}(x, coef = NULL, log = TRUE, ...)
}
\arguments{
\item{x}{Fitted \code{\link{ridgetrace}} model object.}

\item{coef}{A vector of true coefficients if available.}

\item{log}{Boolean. If \code{log = TRUE}, the default, horizontal axis will
display the logarithm of lambda.}

\item{...}{additional arguments.}
}
\value{
Supports are returned.
}
\description{
Plot Diagnostics for a \code{\link{ridgetrace}} Object
}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}

res.ridgetrace <-
  ridgetrace(
    formula = y ~ X001 + X002 + X003 + X004 + X005,
    data = dataGCE)

plot(res.ridgetrace)
}
