% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.lmgceclassfunctions.R
\name{print.cv.lmgce}
\alias{print.cv.lmgce}
\title{Print \code{\link{cv.lmgce}} object}
\usage{
\method{print}{cv.lmgce}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{fitted \code{\link{cv.lmgce}} object.}

\item{digits}{significant digits in printout.}

\item{...}{additional print arguments.}
}
\value{
A small summary of a \code{cv.lmgce} object is returned.
}
\description{
Print \code{\link{cv.lmgce}} object
}
\examples{
\donttest{
res.cv.lmgce <-
  cv.lmgce(y ~ .,
           data = dataGCE)

res.cv.lmgce
}

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
