% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.tsbootgceclassfunctions.R
\name{print.cv.tsbootgce}
\alias{print.cv.tsbootgce}
\title{Print \code{\link{cv.tsbootgce}} object}
\usage{
\method{print}{cv.tsbootgce}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{fitted \code{cv.tsbootgce} object.}

\item{digits}{significant digits in printout.}

\item{...}{additional print arguments.}
}
\value{
A small summary of a \code{\link{tsbootgce}} object is returned.
}
\description{
Print \code{\link{cv.tsbootgce}} object
}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
