% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmgceclassfunctions.R
\name{print.lmgce}
\alias{print.lmgce}
\title{Print a \code{\link{lmgce}} object}
\usage{
\method{print}{lmgce}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{fitted \code{\link{lmgce}} object.}

\item{digits}{significant digits in printout.}

\item{...}{additional print arguments.}
}
\value{
A small summary of a \code{\link{lmgce}} object is returned.
}
\description{
Concise summary of a \code{\link{lmgce}} object
}
\examples{
\donttest{
res_gce_package <-
  lmgce(y ~ .,
        data = dataGCE,
        boot.B = 50,
        seed = 230676)
}
res_gce_package

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
