% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgetraceclassfunctions.R
\name{print.ridgetrace}
\alias{print.ridgetrace}
\title{Print a \code{\link{ridgetrace}} object}
\usage{
\method{print}{ridgetrace}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{fitted \code{\link{ridgetrace}} object.}

\item{digits}{significant digits in printout.}

\item{...}{additional print arguments.}
}
\value{
A small summary of a \code{\link{ridgetrace}} object is returned.
}
\description{
Concise summary of a \code{\link{ridgetrace}} object
}
\examples{
res.ridgetrace <-
  ridgetrace(
    formula = y ~ X001 + X002 + X003 + X004 + X005,
    data = dataGCE)

res.ridgetrace

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
