% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbootgceclassfunctions.R
\name{print.tsbootgce}
\alias{print.tsbootgce}
\title{Print \code{\link{tsbootgce}} object}
\usage{
\method{print}{tsbootgce}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{fitted \code{tsbootgce} object.}

\item{digits}{significant digits in printout.}

\item{...}{additional print arguments.}
}
\value{
A small summary of a \code{\link{tsbootgce}} object is returned.
}
\description{
Print \code{\link{tsbootgce}} object
}
\examples{
\donttest{
res.tsbootgce <-
  tsbootgce(
    formula = CO2 ~ 1 + L(EPC, 1) + L(EUS, 2) + L(GDP, 0),
    data = moz_ts)

res.tsbootgce
}

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
