% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_gce_package.R
\docType{data}
\name{res_gce_package}
\alias{res_gce_package}
\title{Example `lmgce` object}
\format{
An object of class `"lmgce"`.
}
\source{
generated by the package.
}
\usage{
res_gce_package
}
\description{
An example of an object of class `lmgce` used for demonstration.
}
\keyword{datasets}
