% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalebackcoef.R
\name{scalebackcoef}
\alias{scalebackcoef}
\title{Scale coefficients back}
\usage{
scalebackcoef(X.scaled, y.scaled, betas.scaled, intercept = TRUE)
}
\arguments{
\item{X.scaled}{A matrix scaled with \code{\link[base]{scale}}.}

\item{y.scaled}{A vector scaled with \code{\link[base]{scale}}.}

\item{betas.scaled}{A vector of given scaled coefficients.}

\item{intercept}{logical indicating if intercept is to be calculated}
}
\value{
Returns a vector of unscaled numeric regression coefficients.
}
\description{
Given a vector of scaled (standardized) regression coefficients the function
returns the unscaled (in the original scale) regression coefficients
}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
