% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmgceclassfunctions.R
\name{variable.names.lmgce}
\alias{variable.names.lmgce}
\title{Variable Names of \code{\link{lmgce}} Fitted Models}
\usage{
\method{variable.names}{lmgce}(object, ...)
}
\arguments{
\item{object}{Fitted \code{\link{lmgce}} model object.}

\item{...}{additional arguments.}
}
\value{
A character vector containing the names of the variables in the
  \code{\link{lmgce}} model object.
}
\description{
Simple utility returning variable names.
}
\examples{
\donttest{
res_gce_package <-
  lmgce(y ~ .,
        data = dataGCE,
        boot.B = 50,
        seed = 230676)
}

variable.names(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
