% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmgceclassfunctions.R
\name{case.names.lmgce}
\alias{case.names.lmgce}
\title{Case Names of \code{\link{lmgce}} Fitted Models}
\usage{
\method{case.names}{lmgce}(object, ...)
}
\arguments{
\item{object}{Fitted \code{\link{lmgce}} model object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character vector containing the names or labels of the cases
(observations) in the \code{\link{lmgce}} model object.
}
\description{
Simple utility returning case names.
}
\examples{
\donttest{
res_gce_package <-
  lmgce(y ~ .,
        data = dataGCE,
        boot.B = 50,
        seed = 230676)
}
case.names(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
